requirements.txt
fastapi==0.110.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
bittensor>=7.0.0
bittensor-wallet>=1.0.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0
requests>=2.31.0
tweepy>=4.19.0